backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.4
bcrypt==4.0.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
# mediapipe>=0.10.0  # Commented out for now - install separately if needed
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
python-dateutil==2.8.2
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
